refactor: Use prisma types for Task and TaskState
