+++ services/task-workflow/src/domains/task/graphql-resolvers/query.ts
@@ -1,21 +1,12 @@
-export const taskQueryResolver = {
+import { PrismaClient, Task } from '@prisma/client';
+
+export const taskQueryResolvers = {
     Query: {
-        tasks: () => tasks,
+        tasks: getTasks,
     },
 };
 
-const tasks = [
-    {
-        id: 'task-1',
-        name: 'Create GQL Mutations',
-        description: 'Create GQL Mutations for task creation',
-        state: 'Planned',
-    },
-    {
-        id: 'task-2',
-        name: 'Implement Unit Tests',
-        description: 'Implement unit tests for GQL resolvers',
-        state: 'Planned',
-        comment: 'depends on task-1',
-    },
-];
+async function getTasks(): Promise<Task[]> {
+    const prismaClient = new PrismaClient();
+    return prismaClient.task.findMany();
+}
