+++ services/task-workflow/src/domains/task/index.ts
@@ -1,4 +1,4 @@
-export * from './task';
+export * from './task-events';
 export * from './task-command-utils';
 export * from './task-query-utils';
 export * from './graphql-resolvers';
