+++ services/task-workflow/src/domains/task/task-events.ts
@@ -1,0 +1,10 @@
+/**
+ * All domain events related to Task.
+ */
+export enum TaskEvents {
+    Create = 'Create',
+    MoveToInProgress = 'MoveToInProgress',
+    MoveToCompleted = 'MoveToCompleted',
+    MoveToPaused = 'MoveToPaused',
+    Abandon = 'Abandon',
+}
